Move fft functions to fft file
